// packages/frontend/src/components/workflow-run-debugger/components/panels/inspector-panel/overview.types.ts
/*
 * Hexabot — Fair Core License (FCL-1.0-ALv2)
 * Copyright (c) 2026 Hexastack.
 * Full terms: see LICENSE.md.
 */

export type OverviewLabels = {
  none: string;
  noData: string;
  name: string;
  status: string;
  triggeredAt: string;
  triggeredBy: string;
  duration: string;
  error: string;
  input: string;
  context: string;
  output: string;
};